[Chore] 주석, 문서 수정
