+++ docs/REQUIREMENTS.md
@@ -51,10 +51,10 @@
 ## 웹 기반
 
 - [ ] 웹 UI 를 사용해 로또 게임의 주요 기능을 사용할 수 있어야 한다.
-  - [ ] 로또 구매
-  - [ ] 구매한 로또 목록 확인
-  - [ ] 당첨 번호 및 보너스 번호 지정
-  - [ ] 당첨 통계 확인
+  - [x] 로또 구매
+  - [x] 구매한 로또 목록 확인
+  - [x] 당첨 번호 및 보너스 번호 지정
+  - [x] 당첨 통계 확인
   - [ ] 게임 재시작
 
 ## 콘솔 기반
