+++ index.html
@@ -139,7 +139,9 @@
           </div>
           <p class="text-center font-bold" id="profit-rate"></p>
           <div class="d-flex justify-center mt-5">
-            <button type="button" class="btn btn-cyan">다시 시작하기</button>
+            <button type="button" class="btn btn-cyan retry-button">
+              다시 시작하기
+            </button>
           </div>
         </div>
       </div>
