+++ src/js/utils/withRetry.js
@@ -7,6 +7,7 @@ export const withRetry = async (action) => {
       flag = false
     } catch (error) {
       flag = true
+      // FIXME: 에러를 처리할 수 있는 별도 객체에 전달해주자
       console.log(error.message)
     }
   }
